refactor: 브랜드 테스트 코드 위치 변경
